requirements.txt
fastapi
uvicorn[standard]
anthropic
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.2.1
pydantic
